+++ src/Core/SFX/audioRepository.java
@@ -1,13 +1,11 @@
 package Core.SFX;
 
-import java.net.URL;
-import java.util.Random;
-
 /**
  * KM
  * May 18 2017
  * Repository for storing audio file declarations to ease the stress off of the program later on.
  */
+
 public class audioRepository {
 
 
@@ -16,62 +14,9 @@ public class audioRepository {
     public static int ambianceVolume = 15;
     public static int voiceVolume = 70;
     public static int effectsVolume = 50;
-    public static audioCore music;
-    public static audioCore ambiance;
 
-    //BROKEN
-    /* public static void musicPlay(String toPlay, boolean isShuffle, boolean loop) {
-
-
-        //TODO: Improve efficiency whenever.
-
-        if (!toPlay.equals("")) {
-            music = new audioCore(toPlay, musicVolume, isShuffle, loop); //loads new audio
-            music.start();
-        } else {
-
-            Random shuffle = new Random();
-            int musicPlay = 1 + shuffle.nextInt(2);
-
-            switch (musicPlay) { //BEWARE THE DREADED CHEESE CODING
-                case 1:
-                    music.loadMusic("imperial_fleet.mp3", musicVolume, true, false);
-                    break;
-                case 2:
-                    music.loadMusic("new_dawn.mp3", musicVolume, true, false);
-                    break;
-                case 3:
-                    music.loadMusic("in_search_of_life.mp3", musicVolume, true, false);
-                    break;
-                case 4:
-                    music.loadMusic("mercedes_romero.mp3", musicVolume, true, false);
-                    break;
-                case 5:
-                    music.loadMusic("spatial_lullaby.mp3", musicVolume, true, false);
-                    break;
-                case 6:
-                    music.loadMusic("towards_utopia.mp3", musicVolume, true, false);
-                    break;
-                case 7:
-                    music.loadMusic("to_the_ends_of_the_galaxy.mp3", musicVolume, true, false);
-                    break;
-                case 8:
-                    music.loadMusic("gravitational_constant.mp3", musicVolume, true, false);
-                    break;
-                case 9:
-                    music.loadMusic("assembling_the_fleet.mp3", musicVolume, true, false);
-                    break;
-                default: //fallback in case something bugs out
-                    music.loadMusic("new_dawn.mp3", musicVolume, true, false);
-                    break;
-            }
-
-
-
-        }
-
-    }
-    */
+    private static audioCore music;
+    private static audioCore ambiance;
 
     /** UI actions **/
 
@@ -107,8 +52,9 @@ public class audioRepository {
     }
 
     public static void musicTitleScreen() { //plays the title screen music
+        music.stopAudio();
         music.interrupt();
-        music = new audioCore("/music/creation_and_beyond.mp3", musicVolume);
+        music = new audioCore("/music/creation_and_beyond.mp3", musicVolume, true);
         music.start();
     }
 
@@ -117,13 +63,18 @@ public class audioRepository {
         music.start();
     }
 
-    public static void musicMainGame() {
+    public static void musicShuffle() {
+        music.stopAudio();
         music.interrupt();
-        music = new audioCore("/music/to_the_ends_of_the_galaxy.mp3", musicVolume);
+        music = new audioCore(musicVolume, true);
         music.start();
     }
 
     /** Ambiance **/
+
+    public static void setAmbianceVolume() {
+        ambiance.setVolume(ambianceVolume);
+    }
 
     public static void ambianceMainGame() {
         ambiance = new audioCore("/ambiance/space_ambient01.wav", audioRepository.ambianceVolume, true);