+++ src/Core/eventCore.java
@@ -1,0 +1,28 @@
+package Core;
+
+/**
+ * KM
+ * May 24 2017
+ * Handles storage of in-game events.
+ */
+public class eventCore {
+
+    double fundMultiplier; //multiplier applied to player's cash income
+    double resourceMultiplier; //multiplier applied to player's resource income
+    double unrestMultiplier; //multiplier applied to player's global unrest
+    int eventTime; //how long the event lasts for (in turns)
+
+    public eventCore() {
+    }
+
+
+
+
+
+
+
+
+
+
+
+}